Remove AbstractTypeIdResolver.java
