--- events/api/src/main/java/org/projectnessie/events/api/json/AbstractTypeIdResolver.java
@@ -1,52 +1,0 @@
-/*
- * Copyright (C) 2023 Dremio
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.projectnessie.events.api.json;
-
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.databind.DatabindContext;
-import com.fasterxml.jackson.databind.JavaType;
-import com.fasterxml.jackson.databind.jsontype.impl.TypeIdResolverBase;
-
-abstract class AbstractTypeIdResolver extends TypeIdResolverBase {
-
-  private JavaType baseType;
-
-  public AbstractTypeIdResolver() {}
-
-  @Override
-  public void init(JavaType bt) {
-    baseType = bt;
-  }
-
-  @Override
-  public JsonTypeInfo.Id getMechanism() {
-    return JsonTypeInfo.Id.CUSTOM;
-  }
-
-  @Override
-  public String idFromValueAndType(Object value, Class<?> suggestedType) {
-    return idFromValue(value);
-  }
-
-  @Override
-  public JavaType typeFromId(DatabindContext context, String id) {
-    Class<?> subtype = getSubtype(id);
-    assert baseType.getRawClass().isAssignableFrom(subtype);
-    return context.constructSpecializedType(baseType, subtype);
-  }
-
-  protected abstract Class<?> getSubtype(String id);
-}
